% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneTrack.R
\name{addGenePlot}
\alias{addGenePlot}
\title{add gene track to plot}
\usage{
addGenePlot(
  plot,
  txdb,
  chr,
  start_pos,
  end_pos,
  xlab = "",
  gene_symbol_oi = NULL,
  gene_id_oi = NULL,
  sample_n_gene = NULL,
  palette = NULL,
  position = "bottom",
  height = 1,
  OrgDb = NULL,
  show_legend = FALSE
)
}
\arguments{
\item{plot}{ggplot2 object, gene track will be added to it.}

\item{txdb}{txdb object, providing gene annotation.}

\item{chr}{chromosome id.}

\item{start_pos}{start coordiante of windows.}

\item{end_pos}{end coordiante of windows.}

\item{xlab}{x lab.}

\item{gene_symbol_oi}{gene of interest to show in the form of gene symbol.}

\item{gene_id_oi}{gene of interest to show in the form of gene id}

\item{sample_n_gene}{sample N gene to show if there are too many genes. 
This parameter is mutually exclusive with the gene_symbol_oi and gene_id_oi parameter.}

\item{palette}{palette, default "Set3".}

\item{position}{position of gene track be added to("bottom", "top").}

\item{height}{the relative height of gene track(default: 0.5).}

\item{OrgDb}{OrgDb for change gene id to gene symbol.}

\item{show_legend}{show legend or not.}
}
\value{
ggplot object
}
\description{
add gene track to plot
}
