% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_upstream_and_downstream}
\alias{check_upstream_and_downstream}
\title{check upstream and downstream parameter}
\usage{
check_upstream_and_downstream(upstream, downstream)
}
\arguments{
\item{upstream}{upstream}

\item{downstream}{downstream}
}
\description{
check_upstream_and_downstream
}
