% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseeker-package.R
\name{ChIPseekerCache}
\alias{ChIPseekerCache}
\title{Name of the ChIPseeker cache environment (internal static variable)}
\format{
character vector of length 1
}
\usage{
ChIPseekerCache
}
\description{
This variable stores the name of the cache environment used internally
by ChIPseeker to store annotation data (exons, introns, UTRs, etc.) for
efficient repeated access.
}
\keyword{internal}
