% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEO.R
\name{downloadGEObedFiles}
\alias{downloadGEObedFiles}
\title{downloadGEObedFiles}
\usage{
downloadGEObedFiles(genome, destDir = getwd())
}
\arguments{
\item{genome}{genome version}

\item{destDir}{destination folder}
}
\description{
download all BED files of a particular genome version
}
\author{
G Yu
}

