% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeak.R
\name{dropAnno}
\alias{dropAnno}
\title{dropAnno}
\usage{
dropAnno(csAnno, distanceToTSS_cutoff = 10000)
}
\arguments{
\item{csAnno}{output of annotatePeak}

\item{distanceToTSS_cutoff}{distance to TSS cutoff}
}
\value{
csAnno object
}
\description{
dropAnno
}
\details{
drop annotation exceeding distanceToTSS_cutoff
}
\author{
Guangchuang Yu
}

