% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getBioRegionMatrix}
\alias{getBioRegionMatrix}
\title{calculate the BioRegionMatrix by binning}
\usage{
getBioRegionMatrix(
  peak,
  weightCol = NULL,
  windows,
  box = 800,
  min_body_length = 1000,
  upstream = NULL,
  downstream = NULL
)
}
\arguments{
\item{peak}{peak peak file or GRanges object}

\item{weightCol}{weightCol column name of weight, default is NULL}

\item{windows}{windows a collection of region with equal or not equal size, eg. promoter region, gene region.}

\item{box}{the amount of boxes needed to be splited and it should not be more than min_body_length}

\item{min_body_length}{the minimum length that each gene region should be}

\item{upstream}{rel object reflects the percentage of flank extension, e.g rel(0.2)
integer reflects the actual length of flank extension
default(NULL) reflects the gene body with no extension}

\item{downstream}{rel object reflects the percentage of flank extension, e.g rel(0.2)
integer reflects the actual length of flank extension
default(NULL) reflects the gene body with no extension}
}
\value{
bioregionmatrix
}
\description{
Title getBioRegionMatrix
}
