% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getEndRegion}
\alias{getEndRegion}
\title{getEndRegion}
\usage{
getEndRegion(TxDb = NULL, upstream = 1000, downstream = 1000, by = "gene")
}
\arguments{
\item{TxDb}{TxDb}

\item{upstream}{upstream from start site}

\item{downstream}{downstream from start site}

\item{by}{one of 'gene', 'transcript', 'exon', 'intron' , '3UTR' , '5UTR'}
}
\value{
GRanges object
}
\description{
prepare a region center on end site of selected feature
}
