% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getGeneBody}
\alias{getGeneBody}
\title{prepare the genebody region}
\usage{
getGeneBody(TxDb = NULL, type = "genes")
}
\arguments{
\item{TxDb}{TxDb}

\item{type}{by one of "genes", "exon", "intron"}
}
\value{
GRanges object
}
\description{
Title getGeneBody
}
