% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getGenebodyMatrix}
\alias{getGenebodyMatrix}
\title{calculate the genebodymatrix}
\usage{
getGenebodyMatrix(
  peak,
  weightCol = NULL,
  windows,
  box = 800,
  min_body_length = 1000,
  upstream = NULL,
  downstream = NULL,
  ...
)
}
\arguments{
\item{peak}{peak peak file or GRanges object}

\item{weightCol}{weightCol column name of weight, default is NULL}

\item{windows}{windows a collection of region with equal or not equal size, eg. promoter region, gene region.}

\item{box}{the amount of boxes needed to be splited and it should not be more than min_body_length}

\item{min_body_length}{the minimum length that each gene region should be}

\item{upstream}{the percentage of upstream flank extension, e.g 0.2, it means (TSS-20%)}

\item{downstream}{the percentage of downstream flank extension, e.g 0.2, it means (TES+20%)}
}
\value{
bodymatrix
}
\description{
Title getGenebodyMatrix
}
