% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGenomicAnnotation}
\alias{getGenomicAnnotation}
\title{getGenomicAnnotation}
\usage{
getGenomicAnnotation(peaks, distance, tssRegion = c(-3000, 3000), TxDb, level)
}
\arguments{
\item{peaks}{peaks in GRanges object}

\item{distance}{distance of peak to TSS}

\item{tssRegion}{tssRegion, default is -3kb to +3kb}

\item{TxDb}{TxDb object}

\item{level}{one of gene or transcript}
}
\value{
character vector
}
\description{
get Genomic Annotation of peaks
}
\author{
G Yu
}

