% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEO.R
\name{getGEOgenomeVersion}
\alias{getGEOgenomeVersion}
\title{getGEOgenomeVersion}
\usage{
getGEOgenomeVersion()
}
\value{
data.frame
}
\description{
get genome version statistics collecting from GEO ChIPseq data
}
\author{
G Yu
}

