% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getGEOInfo}
\alias{getGEOInfo}
\title{getGEOInfo}
\usage{
getGEOInfo(genome, simplify = TRUE)
}
\arguments{
\item{genome}{genome version}

\item{simplify}{simplify result or not}
}
\value{
data.frame
}
\description{
get subset of GEO information by genome version keyword
}
\author{
G Yu
}

