% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getHeatmapColor}
\alias{getHeatmapColor}
\title{getHeatmapcolor}
\usage{
getHeatmapColor(n)
}
\arguments{
\item{n}{the amount of colors}
}
\value{
String vector contain the name of color schemes
}
\description{
get the gradient color for the heatmap
}
