% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getHeatmapColors}
\alias{getHeatmapColors}
\title{getHeatmapcolors}
\usage{
getHeatmapColors(n)
}
\arguments{
\item{n}{the amount of colors}
}
\value{
String vector contain the name of color schemes
}
\description{
get the gradient colors for the heatmap
}
