% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getTagMatrix.internal}
\alias{getTagMatrix.internal}
\title{getTagMatrix.internal}
\usage{
getTagMatrix.internal(peak, weightCol = NULL, windows, ignore_strand = FALSE)
}
\arguments{
\item{peak}{peak file or GRanges object}

\item{weightCol}{column name of weight, default is NULL}

\item{windows}{a collection of region with equal size, eg. promoter region.}

\item{ignore_strand}{ignore the strand information or not}
}
\value{
tagMatrix
}
\description{
calculate the tag matrix
}
\author{
G Yu
}
