% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getTagMatrix2.binning.internal}
\alias{getTagMatrix2.binning.internal}
\title{internal function}
\usage{
getTagMatrix2.binning.internal(
  peak,
  weightCol = NULL,
  windows,
  windows_name,
  nbin = 800,
  upstream = NULL,
  downstream = NULL,
  ignore_strand = FALSE
)
}
\arguments{
\item{peak}{peak peak file or GRanges object}

\item{weightCol}{column name of weight, default is NULL}

\item{windows}{a collection of region}

\item{windows_name}{the name of windows}

\item{nbin}{the amount of nbines}

\item{upstream}{the distance of upstream extension}

\item{downstream}{the distance of downstream extension}

\item{ignore_strand}{ignore the strand information or not}
}
\description{
internal function
}
