% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getTagMatrix2.internal}
\alias{getTagMatrix2.internal}
\title{getTagMatrix2.internal}
\usage{
getTagMatrix2.internal(
  peak,
  weightCol = NULL,
  windows,
  windows_name,
  ignore_strand = FALSE
)
}
\arguments{
\item{peak}{peak peak file or GRanges object}

\item{weightCol}{column name of weight, default is NULL}

\item{windows}{a collection of region}

\item{windows_name}{the name of windows}

\item{ignore_strand}{ignore the strand information or not}
}
\description{
getTagMatrix2.internal
}
