% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getTagMatrix2}
\alias{getTagMatrix2}
\title{Nested function for getTagMatrix() to deal with multiple windows}
\usage{
getTagMatrix2(
  peak,
  upstream,
  downstream,
  windows_name,
  type,
  by,
  gr = NULL,
  TxDb = NULL,
  weightCol = NULL,
  nbin = NULL,
  verbose = TRUE,
  ignore_strand = FALSE
)
}
\description{
This is an internal function.
}
