% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getTTSRegion}
\alias{getTTSRegion}
\title{getTTSRegion}
\usage{
getTTSRegion(TxDb = NULL, upstream = 1000, downstream = 1000, by = "gene")
}
\arguments{
\item{TxDb}{TxDb}

\item{upstream}{upstream from TSS site}

\item{downstream}{downstream from TSS site}

\item{by}{one of gene or transcript}
}
\value{
GRanges object
}
\description{
get the tts region
}
