% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseeker-package.R
\docType{data}
\name{info}
\alias{info}
\alias{ucsc_release}
\alias{gsminfo}
\alias{tagMatrixList}
\title{Information Datasets}
\description{
Pre-calculated datasets included in the ChIPseeker package:
\itemize{
  \item \code{ucsc_release}: UCSC genome version information
  \item \code{gsminfo}: GEO Sample (GSM) information for ChIP-seq datasets
  \item \code{tagMatrixList}: Pre-calculated tag matrices for example datasets
}
}
\keyword{datasets}
