% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_label}
\alias{make_label}
\title{make label for figures}
\usage{
make_label(type, by)
}
\arguments{
\item{type}{one of "start_site", "end_site", "body"}

\item{by}{one of 'gene', 'transcript', 'exon', 'intron' , '3UTR' , '5UTR', 'UTR'}
}
\description{
make label for figures
}
