% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnno.R
\name{plotAnnoPie.csAnno}
\alias{plotAnnoPie.csAnno}
\title{plotAnnoPie}
\usage{
plotAnnoPie.csAnno(x, ndigit = 2, cex = 0.9, col = NA,
  legend.position = "rightside", pie3D = FALSE, ...)
}
\arguments{
\item{x}{csAnno object}

\item{ndigit}{number of digit to round}

\item{cex}{label cex}

\item{col}{color}

\item{legend.position}{topright or other.}

\item{pie3D}{plot in 3D or not}

\item{...}{extra parameter}
}
\value{
pie plot of peak genomic feature annotation
}
\description{
pieplot from peak genomic annotation
}
\examples{
\dontrun{
require(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
peakfile <- system.file("extdata", "sample_peaks.txt", package="chipseeker")
peakAnno <- annotatePeak(peakfile, TxDb=txdb)
plotAnnoPie(peakAnno)
}
}
\seealso{
\code{\link{annotatePeak}} \code{\link{plotAnnoBar}}
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
