% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotAvgProf.binning}
\alias{plotAvgProf.binning}
\title{plotAvgProf.binning}
\usage{
plotAvgProf.binning(
  tagMatrix,
  xlab = "Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  conf,
  facet = "none",
  free_y = TRUE,
  upstream = NULL,
  downstream = NULL,
  label,
  ...
)
}
\arguments{
\item{tagMatrix}{tagMatrix or a list of tagMatrix}

\item{xlab}{x label}

\item{ylab}{y label}

\item{conf}{confidence interval}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled}

\item{upstream}{rel object reflects the percentage of flank extension, e.g rel(0.2)
integer reflects the actual length of flank extension or TSS region
NULL reflects the gene body with no extension}

\item{downstream}{rel object reflects the percentage of flank extension, e.g rel(0.2)
integer reflects the actual length of flank extension or TSS region
NULL reflects the gene body with no extension}

\item{label}{label}

\item{...}{additional parameter}
}
\value{
ggplot object
}
\description{
plot the profile of peaks  by binning
}
