% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotBioRegion2}
\alias{plotBioRegion2}
\title{plot the profile of peaks that align to bio regions}
\usage{
plotBioRegion2(
  peak,
  weightCol = NULL,
  TxDb = NULL,
  body_type = NULL,
  start_region_by = NULL,
  xlab = "Scaled Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  conf,
  facet = "none",
  free_y = TRUE,
  verbose = TRUE,
  box = 800,
  min_body_length = 1000,
  upstream = NULL,
  downstream = NULL
)
}
\arguments{
\item{peak}{peak file or GRanges object}

\item{weightCol}{weightCol column name of weight}

\item{TxDb}{TxDb object}

\item{body_type}{one of "genes", "exon", "intron"}

\item{start_region_by}{one of "gene", "transcript", "exon", "intron", "3UTR", "5UTR"}

\item{xlab}{x label}

\item{ylab}{y label}

\item{conf}{confidence interval}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled by AvgProf}

\item{verbose}{print message or not}

\item{box}{the amount of boxes needed to be splited and it should not be more than min_body_length}

\item{min_body_length}{the minimum length that each gene region should be}

\item{upstream}{rel object reflects the percentage of flank extension, e.g rel(0.2)
integer reflects the actual length of flank extension or TSS region
NULL reflects the gene body with no extension}

\item{downstream}{rel object reflects the percentage of flank extension, e.g rel(0.2)
integer reflects the actual length of flank extension or TSS region
NULL reflects the gene body with no extension}
}
\value{
ggplot object
}
\description{
Title plotBioRegion2
}
