% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotDistToTSS}
\alias{plotDistToTSS}
\title{plotDistToTSS}
\usage{
plotDistToTSS(peakAnno, distanceColumn = "distanceToTSS", xlab = "",
  ylab = "Binding sites (\%) (5'->3')",
  title = "Distribution of transcription factor-binding loci relative to TSS")
}
\arguments{
\item{peakAnno}{peak annotation}

\item{distanceColumn}{column name of the distance from peak to nearest gene}

\item{xlab}{x label}

\item{ylab}{y lable}

\item{title}{figure title}
}
\value{
bar plot that summarize distance from peak to
TSS of the nearest gene.
}
\description{
plot feature distribution based on the distances to the TSS
}
\examples{
require(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
peakfile <- system.file("extdata", "sample_peaks.txt", package="ChIPseeker")
peakAnno <- annotatePeak(peakfile, TxDb=txdb)
plotDistToTSS(peakAnno)
}
\author{
G Yu
}
\seealso{
\code{\link{annotatePeak}}
}

