% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotGeneBody}
\alias{plotGeneBody}
\title{plot the profile of peaks of the gene regions}
\usage{
plotGeneBody(
  bodymatrix,
  xlab = "Scaled Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  conf,
  facet = "none",
  free_y = TRUE,
  scaledlength = 8000,
  binsize = 50,
  ...
)
}
\arguments{
\item{bodymatrix}{tagMatrix tagMatrix or a list of tagMatrix}

\item{xlab}{x label}

\item{ylab}{y label}

\item{conf}{confidence interval}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled by AvgProf}

\item{scaledlength}{the length of gene regions to be scaled to}

\item{binsize}{the amount of nucleotide base in each box}

\item{...}{}
}
\value{
ggplot object
}
\description{
Title plotGeneBody
}
