% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotMultiProf.normal.internal}
\alias{plotMultiProf.normal.internal}
\title{internal function}
\usage{
plotMultiProf.normal.internal(
  tagMatrix,
  conf,
  xlim = c(-3000, 3000),
  xlab = "Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  facet = "row",
  free_y = TRUE,
  origin_label,
  ...
)
}
\arguments{
\item{tagMatrix}{tagMatrix}

\item{conf}{confidence interval}

\item{xlim}{xlim}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled by AvgProf}

\item{origin_label}{the label of the center}

\item{...}{additional parameter}
}
\description{
internal function
}
