% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotMultiProf.normal}
\alias{plotMultiProf.normal}
\title{internal function}
\usage{
plotMultiProf.normal(
  tagMatrix,
  xlim,
  xlab = "Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  conf,
  facet = "none",
  free_y = TRUE,
  origin_label = "TSS",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{tagMatrix}{tagMatrix}

\item{xlim}{xlim}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{conf}{confidence interval}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled by AvgProf}

\item{origin_label}{the label of the center}

\item{verbose}{print message or not}

\item{...}{additional parameter}
}
\description{
internal function
}
