% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotMultiProf}
\alias{plotMultiProf}
\title{internal function for plotPeakProf3}
\usage{
plotMultiProf(
  tagMatrix,
  conf,
  xlab = "Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  facet = "none",
  free_y = TRUE,
  ...
)
}
\arguments{
\item{tagMatrix}{tagMatrix}

\item{conf}{confidence interval}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled by AvgProf}

\item{...}{additional parameter}
}
\description{
internal function for plotPeakProf3
}
