% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{GRangesList}
\alias{rel}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GenomicRanges}{\code{\link[GenomicRanges:GRangesList]{GRangesList()}}}

  \item{ggplot2}{\code{\link[ggplot2:rel]{rel()}}}
}}

