% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show}
\alias{show}
\alias{show,csAnno,ANY-method}
\title{show method}
\usage{
show(object)
}
\arguments{
\item{object}{A \code{csAnno} instance}
}
\value{
message
}
\description{
show method for \code{csAnno} instance
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

