% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/csAnno.R
\docType{methods}
\name{upsetplot}
\alias{upsetplot}
\title{upsetplot method generics}
\usage{
upsetplot(x, ...)

upsetplot(x, ...)
}
\arguments{
\item{x}{A \code{csAnno} instance}

\item{...}{additional parameter}
}
\value{
plot
}
\description{
upsetplot method generics

upsetplot method generics
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

