\name{addGeneAnno}
\alias{addGeneAnno}
\title{addGeneAnno}
\usage{
addGeneAnno(annoDb, geneID, type)
}
\arguments{
  \item{annoDb}{annotation package}

  \item{geneID}{query geneID}

  \item{type}{gene ID type}
}
\value{
data.frame
}
\description{
add gene annotation, symbol, gene name etc.
}
\author{
G Yu
}

