% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.csAnno}
\alias{as.data.frame.csAnno}
\title{as.data.frame.csAnno}
\usage{
\method{as.data.frame}{csAnno}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{csAnno object}

\item{row.names}{row names}

\item{optional}{should be omitted.}

\item{...}{additional parameters}
}
\value{
data.frame
}
\description{
as.data.frame.csAnno
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

