% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csAnno.R
\name{as.data.frame.csAnno}
\alias{as.data.frame.csAnno}
\title{as.data.frame.csAnno}
\usage{
\method{as.data.frame}{csAnno}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{csAnno object}

\item{row.names}{row names}

\item{optional}{should be omitted.}

\item{...}{additional parameters}
}
\value{
data.frame
}
\description{
convert csAnno object to data.frame
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}

