% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/csAnno.R
\name{as.GRanges}
\alias{as.GRanges}
\title{as.GRanges}
\usage{
as.GRanges(x)
}
\arguments{
\item{x}{csAnno object}
}
\value{
GRanges object
}
\description{
convert csAnno object to GRanges
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

