% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseeker-package.R
\docType{package}
\name{ChIPseeker-package}
\alias{ChIPseeker}
\alias{ChIPseeker-package}
\title{ChIPseeker: ChIPseeker for ChIP peak Annotation, Comparison, and Visualization}
\description{
This package implements functions to retrieve the nearest genes around the peak, annotate genomic region of the peak, statstical methods for estimate the significance of overlap among ChIP peak data sets, and incorporate GEO database for user to compare the own dataset with those deposited in database. The comparison can be used to infer cooperative regulation and thus can be used to generate hypotheses. Several visualization functions are implemented to summarize the coverage of the peak experiment, average profile and heatmap of peaks binding to TSS regions, genomic annotation, distance to TSS, and overlap of peaks or genes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://onlinelibrary.wiley.com/share/author/GYJGUBYCTRMYJFN2JFZZ?target=10.1002/cpz1.585}
  \item Report bugs at \url{https://github.com/YuLab-SMU/ChIPseeker/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

Other contributors:
\itemize{
  \item Ming Li \email{limiang929@gmail.com} [contributor]
  \item Qianwen Wang \email{treywea@gmail.com} [contributor]
  \item Yun Yan \email{youryanyun@gmail.com} [contributor]
  \item Hervé Pagès \email{hpages.on.github@gmail.com} [contributor]
  \item Michael Kluge \email{michael.kluge@bio.ifi.lmu.de} [contributor]
  \item Thomas Schwarzl \email{schwarzl@embl.de} [contributor]
  \item Zhougeng Xu \email{xuzhougeng@163.com} [contributor]
}

}
\keyword{internal}
