% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChIPseeker-package.R
\docType{package}
\name{ChIPseeker-package}
\alias{ChIPseeker}
\alias{ChIPseeker-package}
\title{ChIP-SEQ Annotation, Visualization and Comparison}
\description{
This package is designed for chip-seq data analysis
}
\details{
\tabular{ll}{ Package: \tab ChIPseeker\cr Type: \tab Package\cr Version: \tab
1.5.1\cr Date: \tab 27-04-2015\cr biocViews:\tab ChIPSeq,
Annotation, Software\cr Depends:\tab \cr Imports: \tab methods, ggplot2
\cr Suggests:\tab clusterProfiler, GOSemSim\cr License: \tab
Artistic-2.0\cr }
}
\author{
Guangchuang Yu

Maintainer: Guangchuang Yu \email{guangchuangyu@gmail.com}
}
\keyword{package}

