\docType{package}
\name{ChIPseeker-package}
\alias{ChIPseeker}
\alias{ChIPseeker-package}
\title{ChIP-SEQ Annotation}
\description{
This package is designed for chip-seq data analysis
}
\details{
\tabular{ll}{ Package: \tab ChIPseeker\cr Type: \tab
Package\cr Version: \tab 0.99.0\cr Date: \tab 2-01-2014\cr
biocViews:\tab ChIPSeq, Annotation, Software\cr
Depends:\tab \cr Imports: \tab methods, ggplot2 \cr
Suggests:\tab clusterProfiler, GOSemSim\cr License: \tab
Artistic-2.0\cr }
}
\author{
Guangchuang Yu

Maintainer: Guangchuang Yu \email{guangchuangyu@gmail.com}
}
\keyword{package}

