% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseeker-package.R
\docType{data}
\name{ChIPseekerCache}
\alias{ChIPseekerCache}
\title{Name of the ChIPseeker cache environment (internal static variable)}
\format{
character vector
}
\usage{
ChIPseekerCache
}
\description{
Name of the ChIPseeker cache environment (internal static variable)
}
\keyword{datasets}
