% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covplot.R
\name{covplot}
\alias{covplot}
\title{covplot}
\usage{
covplot(peak, weightCol = NULL, xlab = "Chromosome Size (bp)", ylab = "",
  title = "ChIP Peaks over Chromosomes", chrs = NULL, xlim = NULL)
}
\arguments{
\item{peak}{peak file or GRanges object}

\item{weightCol}{weight column of peak}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{chrs}{selected chromosomes to plot, all chromosomes by default}

\item{xlim}{ranges to plot, default is whole chromosome}
}
\value{
ggplot2 object
}
\description{
plot peak coverage
}
\author{
G Yu
}

