% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covplot.R
\name{covplot}
\alias{covplot}
\title{covplot}
\usage{
covplot(
  peak,
  weightCol = NULL,
  xlab = "Chromosome Size (bp)",
  ylab = "",
  title = "ChIP Peaks over Chromosomes",
  chrs = NULL,
  xlim = NULL,
  lower = 1,
  fill_color = NULL
)
}
\arguments{
\item{peak}{peak file or GRanges object}

\item{weightCol}{weight column of peak}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{chrs}{selected chromosomes to plot, all chromosomes by default}

\item{xlim}{ranges to plot, default is whole chromosome}

\item{lower}{lower cutoff of coverage signal}

\item{fill_color}{specify the color for the plot. Order matters}
}
\value{
ggplot2 object
}
\description{
plot peak coverage
}
\author{
G Yu
}
