% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{csAnno-class}
\alias{csAnno-class}
\alias{plotAnnoBar,csAnno-method}
\alias{plotAnnoPie,csAnno-method}
\alias{plotDistToTSS,csAnno-method}
\alias{show,csAnno-method}
\alias{vennpie,csAnno-method}
\title{Class "csAnno"
This class represents the output of ChIPseeker Annotation}
\description{
Class "csAnno"
This class represents the output of ChIPseeker Annotation
}
\section{Slots}{

\describe{
\item{\code{anno}}{annotation}

\item{\code{tssRegion}}{TSS region}

\item{\code{level}}{transcript or gene}

\item{\code{detailGenomicAnnotation}}{Genomic Annotation in detail}

\item{\code{annoStat}}{annotation statistics}

\item{\code{peakNum}}{number of peaks}
}}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\code{\link{annotatePeak}}
}
\keyword{classes}

