% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.ChIPseekerEnv}
\alias{.ChIPseekerEnv}
\title{env function for ChIPseeker}
\usage{
.ChIPseekerEnv(TxDb, item = "ChIPseekerEnv", force = FALSE)
}
\arguments{
\item{TxDb}{txdb object}

\item{item}{item name}

\item{force}{force to update txdb item in cache or not.}
}
\description{
env function for ChIPseeker
}
