% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GEO.R
\name{downloadGSMbedFiles}
\alias{downloadGSMbedFiles}
\title{downloadGSMbedFiles}
\usage{
downloadGSMbedFiles(GSM, destDir = getwd())
}
\arguments{
\item{GSM}{GSM accession numbers}

\item{destDir}{destination folder}
}
\description{
download BED supplementary files of a list of GSM accession numbers
}
\author{
G Yu
}

