% Generated by roxygen2 (4.0.0): do not edit by hand
\name{enrichAnnoOverlap}
\alias{enrichAnnoOverlap}
\title{enrichAnnoOverlap}
\usage{
enrichAnnoOverlap(queryPeak, targetPeak, TranscriptDb = NULL,
  pAdjustMethod = "BH")
}
\arguments{
\item{queryPeak}{query bed file}

\item{targetPeak}{target bed file(s) or folder containing bed files}

\item{TranscriptDb}{TranscriptDb}

\item{pAdjustMethod}{pvalue adjustment method}
}
\value{
data.frame
}
\description{
calcuate overlap significant of ChIP experiments based on their nearest gene annotation
}
\author{
G Yu
}

