% Generated by roxygen2 (4.0.1): do not edit by hand
\name{enrichAnnoOverlap}
\alias{enrichAnnoOverlap}
\title{enrichAnnoOverlap}
\usage{
enrichAnnoOverlap(queryPeak, targetPeak, TxDb = NULL,
  pAdjustMethod = "BH", chainFile = NULL)
}
\arguments{
\item{queryPeak}{query bed file}

\item{targetPeak}{target bed file(s) or folder containing bed files}

\item{TxDb}{TxDb}

\item{pAdjustMethod}{pvalue adjustment method}

\item{chainFile}{chain file for liftOver}
}
\value{
data.frame
}
\description{
calcuate overlap significant of ChIP experiments based on their nearest gene annotation
}
\author{
G Yu
}

