% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrichOverlap.R
\name{enrichPeakOverlap}
\alias{enrichPeakOverlap}
\title{enrichPeakOverlap}
\usage{
enrichPeakOverlap(queryPeak, targetPeak, TxDb = NULL, pAdjustMethod = "BH",
  nShuffle = 1000, chainFile = NULL, pool = TRUE, ...)
}
\arguments{
\item{queryPeak}{query bed file}

\item{targetPeak}{target bed file(s) or folder that containing bed files}

\item{TxDb}{TxDb}

\item{pAdjustMethod}{pvalue adjustment method}

\item{nShuffle}{shuffle numbers}

\item{chainFile}{chain file for liftOver}

\item{pool}{logical, whether pool target peaks}

\item{...}{additional parameter}
}
\value{
data.frame
}
\description{
calculate overlap significant of ChIP experiments based on the genome coordinations
}
\author{
G Yu
}

