% Generated by roxygen2 (4.0.0): do not edit by hand
\name{enrichPeakOverlap}
\alias{enrichPeakOverlap}
\title{enrichPeakOverlap}
\usage{
enrichPeakOverlap(peak1, peak2, by = "peak", TranscriptDb = NULL,
  pAdjustMethod = "BH", ...)
}
\arguments{
\item{peak1}{peak file in bed format}

\item{peak2}{peak file in bed format}

\item{by}{one of peak or gene}

\item{TranscriptDb}{TranscriptDb}

\item{pAdjustMethod}{pvalue adjust method}

\item{...}{additional parameter}
}
\value{
data.frame
}
\description{
calculate the enrichment of overlap between two peak sets
}
\author{
G Yu
}

