% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichOverlap.R
\name{enrichPeakOverlap}
\alias{enrichPeakOverlap}
\title{enrichPeakOverlap}
\usage{
enrichPeakOverlap(queryPeak, targetPeak, TxDb = NULL,
  pAdjustMethod = "BH", nShuffle = 1000, chainFile = NULL,
  pool = TRUE, mc.cores = detectCores() - 1, verbose = TRUE)
}
\arguments{
\item{queryPeak}{query bed file or GRanges object}

\item{targetPeak}{target bed file(s) or folder that containing bed files or a list of GRanges objects}

\item{TxDb}{TxDb}

\item{pAdjustMethod}{pvalue adjustment method}

\item{nShuffle}{shuffle numbers}

\item{chainFile}{chain file for liftOver}

\item{pool}{logical, whether pool target peaks}

\item{mc.cores}{number of cores, see \link[parallel]{mclapply}}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
calculate overlap significant of ChIP experiments based on the genome coordinations
}
\author{
G Yu
}
