% Generated by roxygen2 (4.0.0): do not edit by hand
\name{enrichPeakOverlap}
\alias{enrichPeakOverlap}
\title{enrichPeakOverlap}
\usage{
enrichPeakOverlap(queryPeak, targetPeak, TranscriptDb = NULL,
  pAdjustMethod = "BH", nShuffle = 1000)
}
\arguments{
\item{queryPeak}{query bed file}

\item{targetPeak}{target bed file(s) or folder that containing bed files}

\item{TranscriptDb}{TranscriptDb}

\item{pAdjustMethod}{pvalue adjustment method}

\item{nShuffle}{shuffle numbers}
}
\value{
data.frame
}
\description{
calculate overlap significant of ChIP experiments based on the genome coordinations
}
\author{
G Yu
}

