% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getBioRegion}
\alias{getBioRegion}
\title{getBioRegion}
\usage{
getBioRegion(
  TxDb = NULL,
  upstream = 1000,
  downstream = 1000,
  by = "gene",
  type = "start_site"
)
}
\arguments{
\item{TxDb}{TxDb}

\item{upstream}{upstream from start site}

\item{downstream}{downstream from start site}

\item{by}{one of 'gene', 'transcript', 'exon', 'intron' , '3UTR' , '5UTR'}

\item{type}{one of "start_site", "end_site", "body"}
}
\value{
GRanges object
}
\description{
prepare a bioregion of selected feature
}
\details{
this function combined previous functions getPromoters(),getBioRegion(),getGeneBody()
https://github.com/GuangchuangYu/ChIPseeker/issues/16
https://github.com/GuangchuangYu/ChIPseeker/issues/87
}
\author{
Guangchuang Yu, Ming L
}
