% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getBioRegion}
\alias{getBioRegion}
\title{getBioRegion}
\usage{
getBioRegion(TxDb = NULL, upstream = 1000, downstream = 1000,
  by = "gene")
}
\arguments{
\item{TxDb}{TxDb}

\item{upstream}{upstream from start site}

\item{downstream}{downstream from start site}

\item{by}{one of 'gene', 'transcript', 'exon', 'intron'}
}
\value{
GRanges object
}
\description{
prepare a region center on start site of selected feature
}
\author{
Guangchuang Yu
}
