% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeneAnno.R
\name{getGeneAnno}
\alias{getGeneAnno}
\title{getGeneAnno}
\usage{
getGeneAnno(annoDb, geneID, type, columns)
}
\arguments{
\item{annoDb}{annotation package}

\item{geneID}{query geneID}

\item{type}{gene ID type}

\item{columns}{names of columns to be obtained from database}
}
\value{
data.frame
}
\description{
get gene annotation, symbol, gene name etc.
}
\author{
G Yu
}
