% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addGeneAnno.R
\name{getGeneAnno}
\alias{getGeneAnno}
\title{getGeneAnno}
\usage{
getGeneAnno(annoDb, geneID, type)
}
\arguments{
\item{annoDb}{annotation package}

\item{geneID}{query geneID}

\item{type}{gene ID type}
}
\value{
data.frame
}
\description{
get gene annotation, symbol, gene name etc.
}
\author{
G Yu
}

