% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getGenebodyMatrix}
\alias{getGenebodyMatrix}
\title{calculate the genebodymatrix}
\usage{
getGenebodyMatrix(
  peak,
  weightCol = NULL,
  windows,
  scaledlength = 8000,
  binsize = 50,
  min_body_length = 1000
)
}
\arguments{
\item{peak}{peak peak file or GRanges object}

\item{weightCol}{weightCol column name of weight, default is NULL}

\item{windows}{windows a collection of region with equal or not equal size, eg. promoter region, gene region.}

\item{scaledlength}{the length that different gene regions are scaled to}

\item{binsize}{the amount of nucleotide base in each box}

\item{min_body_length}{the minimum length that each gene region should be}
}
\value{
bodymatrix
}
\description{
Title getGenebodyMatrix
}
