% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearestFeatureIndicesAndDistances.R
\name{getNearestFeatureIndicesAndDistances}
\alias{getNearestFeatureIndicesAndDistances}
\title{getNearestFeatureIndicesAndDistances}
\usage{
getNearestFeatureIndicesAndDistances(peaks, features, sameStrand = FALSE,
  ignoreOverlap = FALSE, ignoreUpstream = FALSE, ignoreDownstream = FALSE)
}
\arguments{
\item{peaks}{peak in GRanges}

\item{features}{features in GRanges}

\item{sameStrand}{logical, whether find nearest gene in the same strand}

\item{ignoreOverlap}{logical, whether ignore overlap of TSS with peak}

\item{ignoreUpstream}{logical, if True only annotate gene in downstream of peak.}

\item{ignoreDownstream}{logical, if True only annotate gene in upstream of peak.}
}
\value{
list
}
\description{
get index of features that closest to peak and calculate distance
}
\author{
G Yu
}

