% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getNearestFeatureIndicesAndDistances}
\alias{getNearestFeatureIndicesAndDistances}
\title{getNearestFeatureIndicesAndDistances}
\usage{
getNearestFeatureIndicesAndDistances(peaks, features)
}
\arguments{
\item{peaks}{peak in GRanges}

\item{features}{features in GRanges}
}
\value{
list
}
\description{
get index of features that closest to peak and calculate distance
}
\author{
G Yu
}

