\name{getNearestFeatureIndicesAndDistances}
\alias{getNearestFeatureIndicesAndDistances}
\title{getNearestFeatureIndicesAndDistances}
\usage{
getNearestFeatureIndicesAndDistances(peaks, features)
}
\arguments{
  \item{peaks}{peak in GRanges}

  \item{features}{features in GRanges}
}
\value{
data.frame
}
\description{
get index of features that closest to peak and calculate
distance
}
\author{
G Yu
}

