% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPromoters}
\alias{getPromoters}
\title{getPromoters}
\usage{
getPromoters(TranscriptDb = NULL, upstream = 1000, downstream = 1000,
  by = "gene")
}
\arguments{
\item{TranscriptDb}{TranscriptDb}

\item{upstream}{upstream from TSS site}

\item{downstream}{downstream from TSS site}

\item{by}{one of gene or transcript}
}
\value{
GRanges object
}
\description{
prepare the promoter regions
}

