% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getTagMatrix.binning.internal}
\alias{getTagMatrix.binning.internal}
\title{getTagMatrix.binning.internal}
\usage{
getTagMatrix.binning.internal(
  peak,
  weightCol = NULL,
  windows,
  nbin = 800,
  upstream = NULL,
  downstream = NULL
)
}
\arguments{
\item{peak}{peak peak file or GRanges object}

\item{weightCol}{weightCol column name of weight, default is NULL}

\item{windows}{windows a collection of region with equal or not equal size, eg. promoter region, gene region.}

\item{nbin}{the amount of nbines needed to be splited and it should not be more than min_body_length}

\item{upstream}{rel object, NULL or actual number}

\item{downstream}{rel object, NULL or actual number}
}
\value{
tagMatrix
}
\description{
calculate the tagMatrix by binning
the idea was derived from the function of deeptools
https://deeptools.readthedocs.io/en/develop/content/tools/computeMatrix.html
}
