% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagMatrix.R
\name{getTagMatrix}
\alias{getTagMatrix}
\title{getTagMatrix}
\usage{
getTagMatrix(
  peak,
  upstream,
  downstream,
  windows,
  type,
  by,
  TxDb = NULL,
  weightCol = NULL,
  nbin = NULL,
  verbose = TRUE,
  ignore_strand = FALSE
)
}
\arguments{
\item{peak}{peak peak file or GRanges object}

\item{upstream}{the distance of upstream extension}

\item{downstream}{the distance of downstream extension}

\item{windows}{a collection of region}

\item{type}{one of "start_site", "end_site", "body"}

\item{by}{one of 'gene', 'transcript', 'exon', 'intron' , '3UTR' , '5UTR'}

\item{TxDb}{TxDb}

\item{weightCol}{column name of weight, default is NULL}

\item{nbin}{the amount of nbines}

\item{verbose}{print message or not}

\item{ignore_strand}{ignore the strand information or not}
}
\value{
tagMatrix
}
\description{
calculate the tag matrix
}
