% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTagMatrix}
\alias{getTagMatrix}
\title{getTagMatrix}
\usage{
getTagMatrix(peak, weightCol = NULL, windows)
}
\arguments{
\item{peak}{peak file or GRanges object}

\item{weightCol}{column name of weight, default is NULL}

\item{windows}{a collection of region with equal size, eg. promoter region.}
}
\value{
tagMatrix
}
\description{
calculate the tag matrix
}

