% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseeker-package.R
\docType{data}
\name{info}
\alias{info}
\alias{ucsc_release}
\alias{gsminfo}
\alias{tagMatrixList}
\title{Information Datasets}
\description{
ucsc genome version, precalcuated data and gsm information
}
\keyword{datasets}
