% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{overlap}
\alias{overlap}
\title{overlap}
\usage{
overlap(Sets)
}
\arguments{
\item{Sets}{a list of objects}
}
\value{
data.frame
}
\description{
calculate the overlap matrix, which is useful for vennplot
}
\author{
G Yu
}

