% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{peak_Profile_Heatmap}
\alias{peak_Profile_Heatmap}
\title{peak_Profile_Heatmap}
\usage{
peak_Profile_Heatmap(
  peak,
  weightCol = NULL,
  TxDb = NULL,
  upstream = 1000,
  downstream = 1000,
  xlab = "",
  ylab = "",
  title = NULL,
  palette = NULL,
  verbose = TRUE,
  by = "gene",
  type = "start_site",
  nbin = NULL,
  ignore_strand = FALSE,
  windows_name = NULL,
  ncol = NULL,
  nrow = NULL,
  facet_label_text_size = 12,
  conf,
  facet = "row",
  free_y = TRUE,
  height_proportion = 4
)
}
\arguments{
\item{peak}{peak file or GRanges object}

\item{weightCol}{column name of weight}

\item{TxDb}{TxDb object}

\item{upstream}{upstream position}

\item{downstream}{downstream position}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{palette}{palette to be filled in,details see \link[ggplot2]{scale_colour_brewer}}

\item{verbose}{print message or not}

\item{by}{one of 'gene', 'transcript', 'exon', 'intron' , '3UTR' , '5UTR', 'UTR'}

\item{type}{one of "start_site", "end_site", "body"}

\item{nbin}{the amount of nbines}

\item{ignore_strand}{ignore the strand information or not}

\item{windows_name}{the name for each window, which will also be showed in the picture as labels}

\item{ncol}{the ncol of plotting a list of peak}

\item{nrow}{the nrow of plotting a list of peak}

\item{facet_label_text_size}{the size of facet label text}

\item{conf}{confidence interval}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled by AvgProf}

\item{height_proportion}{the proportion of profiling picture and heatmap}
}
\description{
plot peak heatmap and profile in a picture
}
