% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{peakHeatmap_multiple_Sets}
\alias{peakHeatmap_multiple_Sets}
\title{peakHeatmap}
\usage{
peakHeatmap_multiple_Sets(
  peak,
  weightCol = NULL,
  TxDb = NULL,
  upstream = 1000,
  downstream = 1000,
  xlab = "",
  ylab = "",
  title = NULL,
  palette = NULL,
  verbose = TRUE,
  by = "gene",
  type = "start_site",
  nbin = NULL,
  ignore_strand = FALSE,
  windows_name = NULL,
  ncol = NULL,
  nrow = NULL,
  facet_label_text_size = 12
)
}
\arguments{
\item{peak}{peak file or GRanges object}

\item{weightCol}{column name of weight}

\item{TxDb}{TxDb object}

\item{upstream}{upstream position}

\item{downstream}{downstream position}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{palette}{palette to be filled in,details see \link[ggplot2]{scale_colour_brewer}}

\item{verbose}{print message or not}

\item{by}{one of 'gene', 'transcript', 'exon', 'intron' , '3UTR' , '5UTR', 'UTR'}

\item{type}{one of "start_site", "end_site", "body"}

\item{nbin}{the amount of nbines}

\item{ignore_strand}{ignore the strand information or not}

\item{windows_name}{the name for each window, which will also be showed in the picture as labels}

\item{ncol}{the ncol of plotting a list of peak}

\item{nrow}{the nrow of plotting a list of peak}

\item{facet_label_text_size}{the size of facet label text}
}
\value{
figure
}
\description{
plot the heatmap of peaks align to a sets of regions
}
