% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/csAnno.R
\docType{methods}
\name{plotAnnoBar}
\alias{plotAnnoBar}
\alias{plotAnnoBar,list-method}
\alias{plotAnnoBar,csAnno,ANY-method}
\title{plotAnnoBar method generics}
\usage{
plotAnnoBar(x, xlab = "", ylab = "Percentage(\%)",
  title = "Feature Distribution", ...)

\S4method{plotAnnoBar}{list}(x, xlab = "", ylab = "Percentage(\%)",
  title = "Feature Distribution", ...)

plotAnnoBar(x, xlab="", ylab='Percentage(\%)',title="Feature Distribution", ...)
}
\arguments{
\item{x}{\code{csAnno} instance}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{...}{additional paramter}
}
\value{
plot
}
\description{
plotAnnoBar method for \code{csAnno} instance
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
