% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/csAnno.R
\docType{methods}
\name{plotAnnoPie}
\alias{plotAnnoPie}
\alias{plotAnnoPie,csAnno,ANY-method}
\title{plotAnnoPie method generics}
\usage{
plotAnnoPie(x, ndigit = 2, cex = 0.9, col = NA,
  legend.position = "rightside", pie3D = FALSE, ...)

plotAnnoPie(x,ndigit=2, cex=0.9,col=NA,legend.position="rightside", pie3D=FALSE, ...)
}
\arguments{
\item{x}{\code{csAnno} instance}

\item{ndigit}{number of digit to round}

\item{cex}{label cex}

\item{col}{color}

\item{legend.position}{topright or other.}

\item{pie3D}{plot in 3D or not}

\item{...}{extra parameter}
}
\value{
plot
}
\description{
plotAnnoPie method for \code{csAnno} instance
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
