% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotAvgProf}
\alias{plotAvgProf}
\title{plotAvgProf}
\usage{
plotAvgProf(tagMatrix, xlim, xlab = "Genomic Region (5'->3')",
  ylab = "Read Count Frequency")
}
\arguments{
\item{tagMatrix}{tagMatrix or a list of tagMatrix}

\item{xlim}{xlim}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
ggplot object
}
\description{
plot the profile of peaks
}
\author{
G Yu
}

