% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotAvgProf}
\alias{plotAvgProf}
\title{plotAvgProf}
\usage{
plotAvgProf(
  tagMatrix,
  xlim,
  xlab = "Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  conf,
  facet = "none",
  free_y = TRUE,
  ...
)
}
\arguments{
\item{tagMatrix}{tagMatrix or a list of tagMatrix}

\item{xlim}{xlim}

\item{xlab}{x label}

\item{ylab}{y label}

\item{conf}{confidence interval}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled by AvgProf}

\item{...}{additional parameter}
}
\value{
ggplot object
}
\description{
plot the profile of peaks
}
\author{
G Yu; Y Yan
}
