% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/csAnno.R
\docType{methods}
\name{plotDistToTSS}
\alias{plotDistToTSS}
\alias{plotDistToTSS,list-method}
\alias{plotDistToTSS,csAnno,ANY-method}
\title{plotDistToTSS method generics}
\usage{
plotDistToTSS(
  x,
  distanceColumn = "distanceToTSS",
  xlab = "",
  ylab = "Binding sites (\%) (5'->3')",
  title = "Distribution of transcription factor-binding loci relative to TSS",
  ...
)

\S4method{plotDistToTSS}{list}(
  x,
  distanceColumn = "distanceToTSS",
  xlab = "",
  ylab = "Binding sites (\%) (5'->3')",
  title = "Distribution of transcription factor-binding loci relative to TSS",
  distanceBreaks = c(0, 1000, 3000, 5000, 10000, 1e+05),
  palette = NULL,
  ...
)

plotDistToTSS(x,distanceColumn="distanceToTSS", xlab="",
ylab="Binding sites (\%) (5'->3')",
title="Distribution of transcription factor-binding loci relative to TSS",...)
}
\arguments{
\item{x}{\code{csAnno} instance}

\item{distanceColumn}{distance column name}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{...}{additional parameter}

\item{distanceBreaks}{breaks of distance, default is 'c(0, 1000, 3000, 5000, 10000, 100000)'}

\item{palette}{palette name for coloring different distances. Run `RColorBrewer::display.brewer.all()` to see all applicable values.}
}
\value{
plot
}
\description{
plotDistToTSS method for \code{csAnno} instance
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
