% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/csAnno.R
\docType{methods}
\name{plotDistToTSS}
\alias{plotDistToTSS}
\alias{plotDistToTSS,list-method}
\alias{plotDistToTSS,csAnno,ANY-method}
\title{plotDistToTSS method generics}
\usage{
plotDistToTSS(
  x,
  distanceColumn = "distanceToTSS",
  xlab = "",
  ylab = "Binding sites (\%) (5'->3')",
  title = "Distribution of transcription factor-binding loci relative to TSS",
  ...
)

\S4method{plotDistToTSS}{list}(
  x,
  distanceColumn = "distanceToTSS",
  xlab = "",
  ylab = "Binding sites (\%) (5'->3')",
  title = "Distribution of transcription factor-binding loci relative to TSS",
  ...
)

plotDistToTSS(x,distanceColumn="distanceToTSS", xlab="",
ylab="Binding sites (\%) (5'->3')",
title="Distribution of transcription factor-binding loci relative to TSS",...)
}
\arguments{
\item{x}{\code{csAnno} instance}

\item{distanceColumn}{distance column name}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{...}{additional parameter}
}
\value{
plot
}
\description{
plotDistToTSS method for \code{csAnno} instance
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
