% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistToTSS.R
\name{plotDistToTSS.data.frame}
\alias{plotDistToTSS.data.frame}
\title{plotDistToTSS.data.frame}
\usage{
plotDistToTSS.data.frame(
  peakDist,
  distanceColumn = "distanceToTSS",
  xlab = "",
  ylab = "Binding sites (\%) (5'->3')",
  title = "Distribution of transcription factor-binding loci relative to TSS",
  categoryColumn
)
}
\arguments{
\item{peakDist}{peak annotation}

\item{distanceColumn}{column name of the distance from peak to nearest gene}

\item{xlab}{x label}

\item{ylab}{y lable}

\item{title}{figure title}

\item{categoryColumn}{category column}
}
\value{
bar plot that summarize distance from peak to
TSS of the nearest gene.
}
\description{
plot feature distribution based on the distances to the TSS
}
\examples{
\dontrun{
require(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
peakfile <- system.file("extdata", "sample_peaks.txt", package="ChIPseeker")
peakAnno <- annotatePeak(peakfile, TxDb=txdb)
plotDistToTSS(peakAnno)
}
}
\seealso{
\code{\link{annotatePeak}}
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
