% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotGeneBody}
\alias{plotGeneBody}
\title{plot the profile of peaks of the gene regions}
\usage{
plotGeneBody(
  bodymatrix,
  xlab = "Scaled Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  conf,
  facet = "none",
  free_y = TRUE,
  extension = F,
  flankExtension = F,
  flankExtPer = 0.2,
  ...
)
}
\arguments{
\item{bodymatrix}{tagMatrix tagMatrix or a list of tagMatrix}

\item{xlab}{x label}

\item{ylab}{y label}

\item{conf}{confidence interval}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled}

\item{extension}{set TRUE to draw the extension of TSS region}

\item{flankExtension}{set TRUE to extend the flank  of TSS and TES}

\item{flankExtPer}{the percentage of flank extension, e.g 0.2, it means (TSS-20%)~(TES+20%)}

\item{...}{}
}
\value{
ggplot object
}
\description{
Title plotGeneBody
}
