% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotGeneBody2}
\alias{plotGeneBody2}
\title{plot the profile of peaks that align to gene regions}
\usage{
plotGeneBody2(
  peak,
  weightCol = NULL,
  TxDb = NULL,
  type = "genes",
  xlab = "Scaled Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  conf,
  facet = "none",
  free_y = TRUE,
  verbose = TRUE,
  box = 800,
  min_body_length = 1000,
  upstream = 3000,
  downstream = 3000,
  flankExtension = F,
  flankExtPer = 0.2,
  ...
)
}
\arguments{
\item{peak}{peak file or GRanges object}

\item{weightCol}{weightCol column name of weight}

\item{TxDb}{TxDb object}

\item{type}{one of "genes", "exon", "intron", "promoters"}

\item{xlab}{x label}

\item{ylab}{y label}

\item{conf}{confidence interval}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled by AvgProf}

\item{verbose}{print message or not}

\item{box}{the amount of boxes needed to be splited and it should not be more than min_body_length}

\item{min_body_length}{the minimum length that each gene region should be}

\item{upstream}{upstream, when use this function to draw TSS region}

\item{downstream}{downstream, when use this function to draw TSS region}

\item{flankExtension}{set TRUE to extend the flank  of TSS and TES}

\item{flankExtPer}{the percentage of flank extension, e.g 0.2, it means (TSS-20%)~(TES+20%)}
}
\value{
ggplot object
}
\description{
Title plotGeneBody2
}
