% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotPeakProf}
\alias{plotPeakProf}
\title{plotPeakProf}
\usage{
plotPeakProf(
  tagMatrix,
  conf,
  xlab = "Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  facet = "none",
  free_y = TRUE
)
}
\arguments{
\item{tagMatrix}{tagMatrix or a list of tagMatrix}

\item{conf}{confidence interval}

\item{xlab}{x label}

\item{ylab}{y label}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled by AvgProf}
}
\value{
ggplot object
}
\description{
plot the profile of peaks
}
\details{
this function combined previous function plotAvgProf()
}
