% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotPeakProf2}
\alias{plotPeakProf2}
\title{plotAvgProf2}
\usage{
plotPeakProf2(
  peak,
  upstream,
  downstream,
  conf,
  by,
  type,
  weightCol = NULL,
  TxDb = NULL,
  xlab = "Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  facet = "none",
  free_y = TRUE,
  verbose = TRUE,
  nbin = NULL,
  ignore_strand = FALSE,
  ...
)
}
\arguments{
\item{peak}{peak file or GRanges object}

\item{upstream}{upstream position}

\item{downstream}{downstream position}

\item{conf}{confidence interval}

\item{by}{one of 'gene', 'transcript', 'exon', 'intron' , '3UTR' , '5UTR'}

\item{type}{one of "start_site", "end_site", "body"}

\item{weightCol}{column name of weight}

\item{TxDb}{TxDb object}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled by AvgProf}

\item{verbose}{print message or not}

\item{nbin}{the amount of nbines}

\item{ignore_strand}{ignore the strand information or not}

\item{...}{additional parameter}
}
\value{
ggplot object
}
\description{
plot the profile of peaks automatically
}
\author{
G Yu, Ming Li
}
