% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{plotPeakProf2}
\alias{plotPeakProf2}
\title{plotPeakProf2}
\usage{
plotPeakProf2(
  peak,
  upstream,
  downstream,
  conf,
  by,
  type,
  weightCol = NULL,
  TxDb = NULL,
  gr,
  xlab = "Genomic Region (5'->3')",
  ylab = "Peak Count Frequency",
  facet = "none",
  free_y = TRUE,
  verbose = TRUE,
  nbin = NULL,
  ignore_strand = FALSE,
  ...
)
}
\arguments{
\item{peak}{peak file or GRanges object}

\item{upstream}{upstream position}

\item{downstream}{downstream position}

\item{conf}{confidence interval}

\item{by}{one of 'gene', 'transcript', 'exon', 'intron' , '3UTR' , '5UTR'}

\item{type}{one of "start_site", "end_site", "body"}

\item{weightCol}{column name of weight}

\item{TxDb}{TxDb object}

\item{gr}{self-made granges object}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{facet}{one of 'none', 'row' and 'column'}

\item{free_y}{if TRUE, y will be scaled by AvgProf}

\item{verbose}{print message or not}

\item{nbin}{the amount of nbines}

\item{ignore_strand}{ignore the strand information or not}

\item{...}{additional parameter}
}
\value{
ggplot object
}
\description{
plot the profile of peaks automatically
}
\details{
`\code{plotPeakProf2()} will call \code{getTagMatrix()} function.
\code{getTagMatrix()} function can produce the matrix for visualization.
\code{peak} stands for the peak file. \code{window} stands for a collection of regions
that users want to look into. Users can use \code{window} to capture the peak of interest.
There are two ways to input \code{window}. 

The first way is that users can use
\code{getPromoters()/getBioRegion()/makeBioRegionFromGranges()} to get \code{window} and
put it into \code{getTagMatrix()}. 

The second way is that users can use \code{getTagMatrix()} to
call \code{getPromoters()/getBioRegion()/makeBioRegionFromGranges()}. In this way
users do not need to input \code{window} parameter but they need to input
\code{txdb} or \code{gr (self-made granges object)}. 

\code{txdb} is a set of packages contained annotation 
of regions of different genomes. Users can
get the regions of interest through specific functions. These specific functions
are built in \code{getPromoters()/getBioRegion()}. Many regions can not be gain
through \code{txdb}, like insulator and enhancer regions. Users can provide these
regions in the form of granges object. These self-made granges object will be passed
to \code{makeBioRegionFromGranges()} to produce the \code{window}.

Details see \code{\link{getPromoters}},\code{\link{getBioRegion}} and \code{\link{makeBioRegionFromGranges}}

\code{upstream} and \code{downstream} parameter have different usages:

(1) \code{window} parameter is provided, 

if \code{type == 'body'}, \code{upstream} and \code{downstream} can use to extend 
the flank of body region.

if \code{type == 'start_site'/'end_site'}, \code{upstream} and \code{downstream} do not
play a role in \code{getTagMatrix()} function.

(2) \code{window} parameter is missing,

if \code{type == 'body'}, \code{upstream} and \code{downstream} can use to extend 
the flank of body region.

if \code{type == 'start_site'/'end_site'}, \code{upstream} and \code{downstream} refer to
the upstream and downstream of the start_site or the end_site.

\code{weightCol} refers to column in peak file. This column acts as a weight vaule. Details
see \url{https://github.com/YuLab-SMU/ChIPseeker/issues/15}

\code{nbin} refers to the number of bins. \code{getTagMatrix()} provide a binning method
to get the tag matrix.
}
\author{
G Yu, Ming Li
}
