% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readPeakFile}
\alias{readPeakFile}
\title{readPeakFile}
\usage{
readPeakFile(peakfile, as = "GRanges")
}
\arguments{
\item{peakfile}{peak file}

\item{as}{output format, one of GRanges or data.frame}
}
\value{
peak information, in GRanges or data.frame object
}
\description{
read peak file and store in data.frame or GRanges object
}
\examples{
peakfile <- system.file("extdata", "sample_peaks.txt", package="ChIPseeker")
peak.gr <- readPeakFile(peakfile, as="GRanges")
peak.gr
}
\author{
G Yu
}

