% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seq2gene.R
\name{seq2gene}
\alias{seq2gene}
\title{seq2gene}
\usage{
seq2gene(seq, tssRegion, flankDistance, TxDb)
}
\arguments{
\item{seq}{genomic regions in GRanges object}

\item{tssRegion}{TSS region}

\item{flankDistance}{flanking search radius}

\item{TxDb}{TranscriptDb object}
}
\value{
gene vector
}
\description{
annotate genomic regions to genes in many-to-many mapping
}
\author{
Guangchuang Yu
}

