% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichOverlap.R
\name{shuffle}
\alias{shuffle}
\title{shuffle}
\usage{
shuffle(peak.gr, TxDb)
}
\arguments{
\item{peak.gr}{GRanges object}

\item{TxDb}{TxDb}
}
\value{
GRanges object
}
\description{
shuffle the position of peak
}
\author{
G Yu
}
