% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shuffle}
\alias{shuffle}
\title{shuffle}
\usage{
shuffle(peak.gr, TranscriptDb)
}
\arguments{
\item{peak.gr}{GRanges object}

\item{TranscriptDb}{TranscriptDb}
}
\value{
GRanges object
}
\description{
shuffle the position of peak
}
\author{
G Yu
}

