% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tagHeatmap}
\alias{tagHeatmap}
\title{tagHeatmap}
\usage{
tagHeatmap(tagMatrix, xlim, xlab = "", ylab = "", title = NULL, color)
}
\arguments{
\item{tagMatrix}{tagMatrix or a list of tagMatrix}

\item{xlim}{xlim}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{color}{color}
}
\value{
figure
}
\description{
plot the heatmap of tagMatrix
}
\author{
G Yu
}

