% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{tagHeatmap}
\alias{tagHeatmap}
\title{tagHeatmap}
\usage{
tagHeatmap(
  tagMatrix,
  xlim,
  xlab = "",
  ylab = "",
  title = NULL,
  color = "RdYlBu"
)
}
\arguments{
\item{tagMatrix}{tagMatrix or a list of tagMatrix}

\item{xlim}{xlim}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{color}{color like c("RdYlBu","RdYlGn","Spectral","RdGy","RdBu","PuOr","PRGn","PiYG","BrBG")}
}
\value{
figure
}
\description{
plot the heatmap of tagMatrix
}
\author{
G Yu
}
