% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagMatrix.R
\name{tagHeatmap}
\alias{tagHeatmap}
\title{tagHeatmap}
\usage{
tagHeatmap(
  tagMatrix,
  xlim = NULL,
  xlab = "",
  ylab = "",
  title = NULL,
  palette = "RdBu",
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{tagMatrix}{tagMatrix or a list of tagMatrix}

\item{xlim}{xlim}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{palette}{palette to be filled in,details see \link[ggplot2]{scale_colour_brewer}}

\item{nrow}{the nrow of plotting a list of peak}

\item{ncol}{the ncol of plotting a list of peak}
}
\value{
figure
}
\description{
plot the heatmap of tagMatrix
}
\author{
G Yu
}
