% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csAnno.R
\docType{methods}
\name{upsetplot}
\alias{upsetplot}
\title{upsetplot method}
\usage{
upsetplot(x, ...)
}
\arguments{
\item{x}{A \code{csAnno} instance}

\item{...}{additional parameter}
}
\value{
plot
}
\description{
upsetplot method generics
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
