% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/csAnno.R
\docType{methods}
\name{vennpie}
\alias{vennpie}
\title{vennpie method generics}
\usage{
vennpie(x, r = 0.2, ...)

vennpie(x, r=0.2, ...)
}
\arguments{
\item{x}{A \code{csAnno} instance}

\item{r}{initial radius}

\item{...}{additional parameter}
}
\value{
plot
}
\description{
vennpie method generics

vennpie method generics
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

