% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vennplot.peakfile}
\alias{vennplot.peakfile}
\title{vennplot.peakfile}
\usage{
vennplot.peakfile(files, labels = NULL)
}
\arguments{
\item{files}{peak files}

\item{labels}{labels for peak files}
}
\value{
figure
}
\description{
vennplot for peak files
}
\author{
G Yu
}

