% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vennplot.R
\name{vennplot}
\alias{vennplot}
\title{vennplot}
\usage{
vennplot(Sets, by = "gplots")
}
\arguments{
\item{Sets}{a list of object, can be vector or GRanges object}

\item{by}{one of gplots or Vennerable}
}
\value{
venn plot that summarize the overlap of peaks
from different experiments or gene annotation from
different peak files.
}
\description{
plot the overlap of a list of object
}
\examples{
## example not run
## require(TxDb.Hsapiens.UCSC.hg19.knownGene)
## txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
## peakfiles <- getSampleFiles()
## peakAnnoList <- lapply(peakfiles, annotatePeak)
## names(peakAnnoList) <- names(peakfiles)
## genes= lapply(peakAnnoList, function(i) as.data.frame(i)$geneId)
## vennplot(genes)
}
\author{
G Yu
}

